feat(engine): support shorthand notation inside ~H sigil
The full-module notation (`<MyMod.button>`) was already supported in "go
to definition", because when passed to ElixirSense, it correctly
recognized the form and found the appropriate module and the function.
However, with a shorthand notation it was not so simple, because
`.button` is not valid Elixir and ElixirSense was not able to make
anything of it.

This implements the support for shorthand notation by doing a
preliminary step before sending the code to ElixirSense. It modifies the
AST and the Document of ForgeAnalysis and replaces all the calls to
`<.button>` with `< button(assigns)` so that ElixirSense can correctly
interpret it as a local function call with arity 1.

This works for functions defined in the same module, but also for
imported functions. Support for ending tag is also included.

While this might seem as a hacky solution, it makes handling shorthand
notation as close as possible to handling the full-module notation,
making sure these two stay conceptually close.